+++ javaScript/projects/testing-practice/src/testing/capitalize.test.js
@@ -1,0 +1,9 @@
+import capitalize from '../containers/capitalize';
+// test()
+test('J should be capitalized', () => {
+  expect(capitalize('jim')).toMatch(/[A-Z]{1}\w+/);
+});
+
+test('P should be capitalized', () => {
+  expect(capitalize('pam')).toMatch(/[A-Z]{1}\w+/);
+});